数据库课程设计/系统源代码文件/requirements.txt
Flask 
mysql-connector-python
Flask-Cors
python-dotenv
werkzeug
requests
pytest